% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10am.elixhauser-package.R
\docType{package}
\name{icd10am.elixhauser-package}
\alias{icd10am.elixhauser}
\alias{icd10am.elixhauser-package}
\title{icd10am.elixhauser: Elixhauser Comorbidity Index for ICD-10-AM Coded Data}
\description{
Applies the Elixhauser Comorbidity Index to Australian admitted patient care data coded using ICD-10-AM. Supports both wide and long input formats and returns van Walraven weighted scores.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/healthpolicy/icd10am.elixhauser}
  \item Report bugs at \url{https://github.com/healthpolicy/icd10am.elixhauser/issues}
}

}
\author{
\strong{Maintainer}: Jim Pearse \email{ejpearse@gmail.com}

}
\keyword{internal}
